+++ tests/test_dunamai.py
@@ -23,9 +23,9 @@ def chdir(where: Path) -> None:
 
 
 def make_run_callback(where: Path) -> Callable:
-    def inner(command):
+    def inner(command, expected_code=0):
         code, out = _run_cmd(command, where=where)
-        if code != 0:
+        if code != expected_code:
             raise RuntimeError("Got exit code {} from command '{}'. Output: {}".format(code, command, out))
         return out
     return inner
@@ -54,9 +54,6 @@ def test__version__init():
     assert v.commit == "abc"
     assert v.dirty
 
-    with pytest.raises(ValueError):
-        Version("1", pre=("x", 3))
-
 
 def test__version__str():
     v = Version(
@@ -84,7 +81,7 @@ def test__version__ordering():
         Version("0.1.0") < "0.2.0"
 
 
-def test__version__serialize():
+def test__version__serialize__pep440():
     assert Version("0.1.0").serialize() == "0.1.0"
 
     assert Version(
@@ -108,10 +105,38 @@ def test__version__serialize():
     assert Version("1", pre=("rc", 3)).serialize() == "1rc3"
 
 
-def test__version__serialize__with_metadata():
+def test__version__serialize__semver():
+    assert Version("0.1.0").serialize(style="semver") == "0.1.0"
+
+    assert Version(
+        "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True,
+    ).serialize(style="semver") == "0.1.0-epoch.2.alpha.3.post.4.dev.5+abc"
+    assert Version(
+        "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True,
+    ).serialize(with_dirty=True, style="semver") == "0.1.0-epoch.2.alpha.3.post.4.dev.5+abc.dirty"
+    assert Version(
+        "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True,
+    ).serialize(with_metadata=False, style="semver") == "0.1.0-epoch.2.alpha.3.post.4.dev.5"
+    assert Version(
+        "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True,
+    ).serialize(with_metadata=False, with_dirty=True, style="semver") == "0.1.0-epoch.2.alpha.3.post.4.dev.5"
+
+    assert Version(
+        "0.1.0", epoch=0, pre=("alpha", 0), post=0, dev=0, commit="000", dirty=False,
+    ).serialize(style="semver") == "0.1.0-epoch.0.alpha.0.post.0.dev.0+000"
+
+    assert Version("0.1.0", pre=("beta", 3)).serialize(style="semver") == "0.1.0-beta.3"
+    assert Version("0.1.0", pre=("rc", 3)).serialize(style="semver") == "0.1.0-rc.3"
+
+
+def test__version__serialize__pep440_with_metadata():
     assert Version("0.1.0").serialize() == "0.1.0"
     assert Version("0.1.0").serialize(with_metadata=True) == "0.1.0"
     assert Version("0.1.0").serialize(with_metadata=False) == "0.1.0"
+
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize() == "0.1.0a1"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=True) == "0.1.0a1+abc"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=False) == "0.1.0a1"
 
     assert Version("0.1.0", post=1, commit="abc").serialize() == "0.1.0.post1+abc"
     assert Version("0.1.0", post=1, commit="abc").serialize(with_metadata=True) == "0.1.0.post1+abc"
@@ -122,7 +147,25 @@ def test__version__serialize__with_metadata():
     assert Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=False) == "0.1.0.dev1"
 
 
-def test__version__serialize__with_dirty():
+def test__version__serialize__semver_with_metadata():
+    assert Version("0.1.0").serialize(style="semver") == "0.1.0"
+    assert Version("0.1.0").serialize(with_metadata=True, style="semver") == "0.1.0"
+    assert Version("0.1.0").serialize(with_metadata=False, style="semver") == "0.1.0"
+
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(style="semver") == "0.1.0-a.1"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=True, style="semver") == "0.1.0-a.1+abc"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=False, style="semver") == "0.1.0-a.1"
+
+    assert Version("0.1.0", post=1, commit="abc").serialize(style="semver") == "0.1.0-post.1+abc"
+    assert Version("0.1.0", post=1, commit="abc").serialize(with_metadata=True, style="semver") == "0.1.0-post.1+abc"
+    assert Version("0.1.0", post=1, commit="abc").serialize(with_metadata=False, style="semver") == "0.1.0-post.1"
+
+    assert Version("0.1.0", dev=1, commit="abc").serialize(style="semver") == "0.1.0-dev.1+abc"
+    assert Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=True, style="semver") == "0.1.0-dev.1+abc"
+    assert Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=False, style="semver") == "0.1.0-dev.1"
+
+
+def test__version__serialize__pep440_with_dirty():
     assert Version("0.1.0", dirty=True).serialize() == "0.1.0"
     assert Version("0.1.0", dirty=True).serialize(with_dirty=True) == "0.1.0+dirty"
 
@@ -135,8 +178,19 @@ def test__version__serialize__with_dirty():
     assert Version("0.1.0", dirty=True).serialize(with_metadata=False) == "0.1.0"
     assert Version("0.1.0", dirty=True).serialize(with_metadata=False, with_dirty=True) == "0.1.0"
 
-    assert Version("0.1.0", post=1, dirty=True).serialize() == "0.1.0.post1"
-    assert Version("0.1.0", post=1, dirty=True).serialize(with_dirty=True) == "0.1.0.post1+dirty"
+
+def test__version__serialize__semver_with_dirty():
+    assert Version("0.1.0", dirty=True).serialize(style="semver") == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(with_dirty=True, style="semver") == "0.1.0+dirty"
+
+    assert Version("0.1.0", dirty=False).serialize(style="semver") == "0.1.0"
+    assert Version("0.1.0", dirty=False).serialize(with_dirty=True, style="semver") == "0.1.0"
+
+    assert Version("0.1.0", dirty=True).serialize(with_metadata=True, style="semver") == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(with_metadata=True, with_dirty=True, style="semver") == "0.1.0+dirty"
+
+    assert Version("0.1.0", dirty=True).serialize(with_metadata=False, style="semver") == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(with_metadata=False, with_dirty=True, style="semver") == "0.1.0"
 
 
 def test__version__serialize__format():
@@ -146,6 +200,14 @@ def test__version__serialize__format():
         "1", epoch=2, pre=("a", 3), post=4, dev=5, commit="abc", dirty=True,
     ).serialize(format=format) == "2,1,a,3,4,5,abc,dirty"
 
+    assert Version("0.1.0").serialize(format="{base}", style="pep440") == "0.1.0"
+    with pytest.raises(ValueError):
+        Version("0.1.0").serialize(format="v{base}", style="pep440")
+
+    assert Version("0.1.0").serialize(format="{base}", style="semver") == "0.1.0"
+    with pytest.raises(ValueError):
+        Version("0.1.0").serialize(format="v{base}", style="semver")
+
 
 def test__version__serialize__error_conditions():
     with pytest.raises(ValueError):
@@ -154,6 +216,9 @@ def test__version__serialize__error_conditions():
         v = Version("1", pre=("a", 3))
         v.pre_type = "x"
         v.serialize()
+    with pytest.raises(ValueError):
+        v = Version("1")
+        v.serialize(style="unknown")
 
 
 @mock.patch("dunamai._run_cmd")
@@ -304,7 +369,13 @@ def test__version__from_git(tmp_path):
         assert from_vcs() == Version("0.1.0", commit="abc", dirty=False)
         assert run("dunamai from git") == "0.1.0"
         assert run("dunamai from any") == "0.1.0"
+
+        # Additional one-off checks not in other VCS integration tests:
+        assert run(r'dunamai from any --pattern "(?P<base>\d\.\d\.\d)"') == "0.1.0"
         assert run('dunamai from any --format "v{base}"') == "v0.1.0"
+        assert run('dunamai from any --style "semver"') == "0.1.0"
+        assert run('dunamai from any --format "v{base}" --style "semver"', 1) \
+            == "Version 'v0.1.0' does not conform to the Semantic Versioning style"
 
         (vcs / "foo.txt").write_text("bye")
         assert from_vcs() == Version("0.1.0", commit="abc", dirty=True)